connection_pool/requirements.txt
psycopg2-binary 2.9.7